services/segmentation-agent/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
torch>=2.1.0
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
